requirements.txt
Adafruit-Blinka==8.69.0
adafruit-circuitpython-busdevice==5.2.15
adafruit-circuitpython-connectionmanager==3.1.6
adafruit-circuitpython-dht==4.0.10
adafruit-circuitpython-requests==4.1.15
adafruit-circuitpython-typing==1.12.3
Adafruit-PlatformDetect==3.86.0
Adafruit-PureIO==1.1.11
annotated-types==0.7.0
binho-host-adapter==0.1.6
click==8.3.1
colorlog==6.10.1
markdown-it-py==4.0.0
mdurl==0.1.2
orjson==3.11.5
peewee==3.19.0
pydantic==2.12.5
pydantic_core==2.41.5
pyftdi==0.57.1
Pygments==2.19.2
pyserial==3.5
python-dotenv==1.2.1
pyusb==1.3.1
PyYAML==6.0.3
rich==14.2.0
shellingham==1.5.4
sysv_ipc==1.2.0
toml==0.10.2
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
